Fix a bug where no item were yielded
No items were yielded if the input array had a single aggregation,
but multiple items in the aggregation.

This was because the `next` method was not calling `forward` if the
inner iterator ends inside the loop (it was only forwarded if the inner
iterator is ended since the beginning). Thus the item being aggregated
was never yielded.